refactor(audience): split filtered audience into its own type
Extracts a ``RapidataAudienceBase`` holding the operations valid on any
audience (``id``, ``name``, ``filters``, ``assign_job``, ``find_jobs``,
``delete``). ``RapidataAudience`` becomes the dimension-audience subclass
and keeps everything that mutates the qualification pool
(``add_classification_example``, ``add_compare_example``,
``update_filters``, ``update_name``, ``get_examples``, ``filter``).

Adds a slim ``RapidataFilteredAudience`` subclass returned by
``RapidataAudience.filter(...)``. It intentionally has no extra methods:
filtered audiences share the base's qualified pool, so adding examples
or updating recruitment filters on the slice would be meaningless or
actively confusing. Pyright now catches that misuse statically instead
of letting it 500 at runtime.

``RapidataBenchmark.create_leaderboard`` widens ``audience_id`` to
``str | RapidataAudienceBase | None`` so both audience types pass.

No behaviour change for existing dimension-audience callers — the public
methods on ``RapidataAudience`` keep the same names and signatures.
